tidy up and fix some errors
