+++ client/src/main.jsx
@@ -7,8 +7,7 @@ import { BrowserRouter } from "react-router-dom";
 createRoot(document.getElementById("root")).render(
 	<StrictMode>
 		<BrowserRouter>
-		<App />
+			<App />
 		</BrowserRouter>
-		
 	</StrictMode>
 );
