+++ client/src/styles/AIRecipes.css
