+++ client/src/styles/SearchInterface.css
